Fix numpy/albumentations compatibility and runtime bugs
- label_utils: use int64 dtype for label encoder (fixes cross_entropy on Windows)
- image_augmentators: fix GaussNoise params for albumentations 2.x (std_range, normalized)
- arguments + params.json: default aug_backend to albumentations
- metric_test_eval: fix raise f-string -> raise ValueError
- Add requirements.txt and plot_metrics.py

+++ src/label_utils.py
@@ -53,7 +53,7 @@ class IncrementalLabelEncoder:
         return self
 
     def transform(self, labels):
-        return np.array([self._label_to_idx[l] for l in labels])
+        return np.array([self._label_to_idx[l] for l in labels], dtype=np.int64)
 
     def inverse_transform(self, indices):
         return np.array([self._classes[i] for i in indices])
